List count of players before the actual list
Report the number of players before `getRecentPlayers` and
`getTodaysPlayers`. These re-use the topic string, which is adjusted for
2 hours (smarter algorithm for "here" coming soon). `!here` is redundant
in having a comma-separated string followed by the monospace table, but
people shouldn't be using it anyways, and we can fix it later.

Refactor `getRecentPlayers` and `getTodaysPlayers` to return an array of
strings instead of a built string.