Clarify intent on dancer count, don't fail when <=13 dancers on first page
If we had <=13 dancers, we'd loop dancerIndex from 1 to 1 at max (won't
run), 0, or a negative number, which caused the initial data fetch to
fail.

To prevent this, use `Math.min` so that we usually grab 7 dancers, but
we actually grab `dancerCount` dancers if there are fewer than 7
dancers. Name variables and comment appropriately.