Abort pruning if we fail to fetch
If a cab fails `retrieveData`, then throw an error and don't retrieve
data or prune players for the rest of this round.

See how this runs over time. If any cab fails for most of the rounds,
then we won't get any updates, and we'll notice in the channel topics.

Might be a temporary band-aid to #38.