Go back to pruning once after retrieving all data, and updating after pruning
Create singular and plural versions of `updatePlayerList` and
`updateChannelTopics`, so that we can run it without arguments and
default to all locations and channels.

Move the prune and these functions to after data has fetched. We don't
have stability problems any more, but we have problems when top-of-list
players switch cabs.